change return type

+++ src/interfaces/slogo/team04/IView.java
@@ -25,5 +25,5 @@ public interface IView{
 	public void update();
 	public void setPenStyle(String selectedItem);
 	public DoubleProperty getPenWidth();
-	public int setShape(int index);
+	public double setShape(int index);
 }
